requirements.txt
python-socketio
python-socketio[client]
panda3d==1.10.11
eventlet